% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_Rpackages.R
\name{hello_Rpackages}
\alias{hello_Rpackages}
\title{A hello package Function}
\usage{
hello_Rpackages(love = TRUE)
}
\arguments{
\item{love}{Do you want to print hello packages? Defaults to TRUE.}
}
\description{
This function allows you to print "Hello Tao, I am R packages".
}
\examples{
hello_Rpackages()
}
\keyword{hello}
